pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "webtest-mcp-server"
version = "0.1.0"
description = "MCP Server for Web UI automation - run Excel test cases via Playwright, AI-triggered"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = []
keywords = ["webtest", "mcp", "playwright", "automation", "excel", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "mcp[cli]>=1.0.0",
    "playwright>=1.40.0",
    "openpyxl>=3.1.0",
    "pyyaml>=6.0",
    "allure-pytest>=2.13.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.optional-dependencies]
dev = [
    "pytest-cov>=4.0.0",
]

[project.scripts]
webtest-mcp = "webtest_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/webtest_mcp"]

[tool.hatch.build.targets.sdist]
packages = ["src/webtest_mcp"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]
